% Auto-generated: do not edit by hand
\name{''appbreadcrumb}

\alias{''appbreadcrumb}

\title{appbreadcrumb component}

\description{
CoreUI breadcrumb component.
}

\usage{
''appbreadcrumb(children=NULL, id=NULL, className=NULL, appRoutes=NULL, tag=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. The children.}

\item{id}{Character. The ID used to identify this component in Dash callbacks, defaults to `appbreadcrumb`.}

\item{className}{Character. The CSS class name, defaults to `app-breadcrumb`.}

\item{appRoutes}{Logical | numeric | character | named list | unnamed list. A list of dictionaries describing URL path in a human-readable way.
Defaults to `[{ path: '/', name: 'Home' }]`.}

\item{tag}{Character. The HTML tag, defaults to `div`.}
}
