% Auto-generated: do not edit by hand
\name{''appcard}

\alias{''appcard}

\title{appcard component}

\description{

}

\usage{
''appcard(children=NULL, id=NULL, className=NULL, tag=NULL, style=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. }

\item{id}{Character. }

\item{className}{Character. }

\item{tag}{Character. }

\item{style}{Named list. }
}
