% Auto-generated: do not edit by hand
\name{''appfooter}

\alias{''appfooter}

\title{appfooter component}

\description{
CoreUI footer component.
}

\usage{
''appfooter(children=NULL, id=NULL, className=NULL, fixed=NULL, tag=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. The children.}

\item{id}{Character. The ID used to identify this component in Dash callbacks, defaults to `appfooter`.}

\item{className}{Character. The CSS class name, defaults to `app-footer`.}

\item{fixed}{Logical. The fixed flag, defaults to `false`.}

\item{tag}{Character. The HTML tag, defaults to `footer`.}
}
