% Auto-generated: do not edit by hand
\name{''appheader}

\alias{''appheader}

\title{appheader component}

\description{
CoreUI header component.
}

\usage{
''appheader(children=NULL, id=NULL, className=NULL, fixed=NULL, tag=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. The children.}

\item{id}{Character. The ID used to identify this component in Dash callbacks, defaults to `appheader`.}

\item{className}{Character. The CSS class name.}

\item{fixed}{Logical. Wether the header is fixed, defaults to `false`.}

\item{tag}{Character. The HTML tag, defaults to `header`.}
}
