% Auto-generated: do not edit by hand
\name{''appheaderdropdown}

\alias{''appheaderdropdown}

\title{appheaderdropdown component}

\description{
CoreUI header dropdown menu component.
}

\usage{
''appheaderdropdown(children=NULL, id=NULL, direction=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. The children.}

\item{id}{Character. The ID used to identify this component in Dash callbacks, defaults to `appheaderdropdown`.}

\item{direction}{Character. The dropdown direction, defaults to `down`.}
}
