% Auto-generated: do not edit by hand
\name{''applinechart}

\alias{''applinechart}

\title{applinechart component}

\description{

}

\usage{
''applinechart(id=NULL, className=NULL, tag=NULL, style=NULL)
}

\arguments{
\item{id}{Character. }

\item{className}{Character. }

\item{tag}{Character. }

\item{style}{Named list. }
}
