% Auto-generated: do not edit by hand
\name{''appnavbarbrand}

\alias{''appnavbarbrand}

\title{appnavbarbrand component}

\description{
CoreUI navbar brand component.
}

\usage{
''appnavbarbrand(children=NULL, tag=NULL, id=NULL, className=NULL, brand=NULL, full=NULL,
minimized=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. The children, defaults to `<img src width height alt className/>`.}

\item{tag}{Character. The HTML tag, defaults to `a`.}

\item{id}{Character. The ID used to identify this component in Dash callbacks, defaults to `appnavbarbrand`.}

\item{className}{Character. The CSS class name, defaults to `navbar-brand`.}

\item{brand}{Logical | numeric | character | named list | unnamed list. The brand image, given as `{src, width, height, alt, className: 'navbar-brand' }`.}

\item{full}{Logical | numeric | character | named list | unnamed list. The full size brand image, given as `{src, width, height, alt, className: 'navbar-brand-full' }`.}

\item{minimized}{Logical | numeric | character | named list | unnamed list. The minimized brand image, given as `{src, width, height, alt, className: 'navbar-brand-minimized' }`.}
}
