% Auto-generated: do not edit by hand
\name{''approuteconditional}

\alias{''approuteconditional}

\title{approuteconditional component}

\description{
Route conditional component.  This component (and its children) are only visible if the browser's current URL matches its route prop. This provides a light-weight client-side alternative to the server-side routing described in https://dash.plot.ly/urls.
}

\usage{
''approuteconditional(children=NULL, id=NULL, route=NULL, className=NULL, tag=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. The children.}

\item{id}{Character. The ID used to identify this component in Dash callbacks, defaults to `approuteconditional`.}

\item{route}{Character. The route. This component is only visible if the browser's current URL matches the contents of this prop.}

\item{className}{Character. The CSS class name, defaults to `route-conditional`.}

\item{tag}{Character. The HTML tag, defaults to `div`.}
}
