% Auto-generated: do not edit by hand
\name{''appsidebar}

\alias{''appsidebar}

\title{appsidebar component}

\description{
CoreUI sidebar component.
}

\usage{
''appsidebar(children=NULL, id=NULL, className=NULL, compact=NULL, display=NULL, fixed=NULL,
minimized=NULL, isOpen=NULL, offCanvas=NULL, staticContext=NULL, tag=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. The children, supports `appsidebarfooter`, `appsidebarform`, `appsidebarheader`, `appsidebarminimizer`, and `appsidebarnav`.}

\item{id}{Character. The ID used to identify this component in Dash callbacks, defaults to `appsidebar`.}

\item{className}{Character. The CSS class name, defaults to `sidebar`.}

\item{compact}{Logical. The compact flag, defaults to `false`.}

\item{display}{Character. The display bootstrap class.}

\item{fixed}{Logical. The fixed flag, defaults to `false`.}

\item{minimized}{Logical. The minimized flag, defaults to `false`.}

\item{isOpen}{Logical. The isOpen flag, defaults to `false`.}

\item{offCanvas}{Logical. The offCanvas flag, defaults to `false`.}

\item{staticContext}{Logical | numeric | character | named list | unnamed list. TODO document this.}

\item{tag}{Character. The HTML tag, defaults to `div`.}
}
