% Auto-generated: do not edit by hand
\name{''appsidebarfooter}

\alias{''appsidebarfooter}

\title{appsidebarfooter component}

\description{
CoreUI sidebar footer component.
}

\usage{
''appsidebarfooter(children=NULL, id=NULL, className=NULL, tag=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. The children.}

\item{id}{Character. The ID used to identify this component in Dash callbacks, defaults to `appsidebarfooter`.}

\item{className}{Character. The CSS class name, defaults to `sidebar-footer`.}

\item{tag}{Character. The HTML tag, defaults to `div`.}
}
