% Auto-generated: do not edit by hand
\name{''appsidebarform}

\alias{''appsidebarform}

\title{appsidebarform component}

\description{
CoreUI sidebar form component.
}

\usage{
''appsidebarform(children=NULL, id=NULL, className=NULL, tag=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. The children.}

\item{id}{Character. The ID used to identify this component in Dash callbacks, defaults to `appsidebarform`.}

\item{className}{Character. The CSS class name, defaults to `sidebar-form`.}

\item{tag}{Character. The HTML tag, defaults to `div`.}
}
