% Auto-generated: do not edit by hand
\name{''appsidebarheader}

\alias{''appsidebarheader}

\title{appsidebarheader component}

\description{
CoreUI sidebar header component.
}

\usage{
''appsidebarheader(children=NULL, id=NULL, className=NULL, tag=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. The children.}

\item{id}{Character. The ID used to identify this component in Dash callbacks, defaults to `appsidebarheader`.}

\item{className}{Character. The CSS class name, defaults to `sidebar-header`.}

\item{tag}{Character. The HTML tag, defaults to `div`.}
}
