% Auto-generated: do not edit by hand
\name{''appsidebartoggler}

\alias{''appsidebartoggler}

\title{appsidebartoggler component}

\description{
CoreUI sidebar toggler component.
}

\usage{
''appsidebartoggler(children=NULL, id=NULL, className=NULL, display=NULL, mobile=NULL, tag=NULL,
type=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. The children, defaults to  `<span className="navbar-toggler-icon" />`.}

\item{id}{Character. The ID used to identify this component in Dash callbacks, defaults to `appsidebartoggler`.}

\item{className}{Character. The CSS class name, defaults to `navbar-toggler`.}

\item{display}{Logical | numeric | character | named list | unnamed list. The display bootstrap class, defaults to `lg`.}

\item{mobile}{Logical. The mobile mode flag, default to `false`.}

\item{tag}{Character. The HTML tag, defaults to `button`.}

\item{type}{Character. The HTML type, defauls to `button`.}
}
