% Auto-generated: do not edit by hand
\name{''apptable}

\alias{''apptable}

\title{apptable component}

\description{

}

\usage{
''apptable(children=NULL, id=NULL, className=NULL, tag=NULL, style=NULL, data=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. }

\item{id}{Character. }

\item{className}{Character. }

\item{tag}{Character. }

\item{style}{Named list. }

\item{data}{Named list. }
}
