% Auto-generated: do not edit by hand
\name{''apptablebody}

\alias{''apptablebody}

\title{apptablebody component}

\description{

}

\usage{
''apptablebody(children=NULL, id=NULL, className=NULL, tag=NULL, style=NULL, data=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. }

\item{id}{Character. }

\item{className}{Character. }

\item{tag}{Character. }

\item{style}{Named list. }

\item{data}{Logical | numeric | character | named list | unnamed list. }
}
